Finished most of the integrations between one controller and other to acheive all functionalities need to add get all users function thats all

+++ Controller/plotController.js
@@ -1,7 +1,13 @@
 import "express-async-errors";
 import { StatusCodes } from "http-status-codes";
 import plotModel from "../models/plotModel.js";
-import { createSlot, updateSlotStatus } from "./slotController.js";
+import {
+  createSlot,
+  deleteAllSlots,
+  isVacant,
+  updateSlotStatus,
+} from "./slotController.js";
+import { BadRequestError } from "../Errors/customErrors.js";
 
 export const getAllPlots = async (req, res) => {
   const plots = await plotModel.find({});
@@ -9,10 +15,21 @@ export const getAllPlots = async (req, res) => {
 };
 
 export const createPlot = async (req, res) => {
-  const plot = await plotModel.create(req.body);
-  res
-    .status(StatusCodes.CREATED)
-    .json({ message: "Plot Created", plot, slot, updatedSlot });
+  const { primeLocationName, address, pinCode, pricePerUnit, numUnits } =
+    req.body;
+  const plot = await plotModel.create({
+    primeLocationName,
+    address,
+    pinCode,
+    pricePerUnit: +pricePerUnit,
+    numUnits: +numUnits,
+    numOccupied: 0,
+    numVacant: +numUnits,
+  });
+  for (let i = 0; i < +numUnits; i++) {
+    await createSlot(plot._id);
+  }
+  res.status(StatusCodes.CREATED).json({ message: "Plot Created", plot });
 };
 
 export const getPlotById = async (req, res) => {
@@ -21,9 +38,15 @@ export const getPlotById = async (req, res) => {
 };
 
 export const updatePlotById = async (req, res) => {
+  const { primeLocationName, address, pinCode, pricePerUnit } = req.body;
   const updatedPlot = await plotModel.findByIdAndUpdate(
     req.params.id,
-    req.body,
+    {
+      primeLocationName,
+      address,
+      pinCode,
+      pricePerUnit: +pricePerUnit,
+    },
     { new: true }
   );
   res
@@ -32,7 +55,10 @@ export const updatePlotById = async (req, res) => {
 };
 
 export const deletePlotById = async (req, res) => {
+  const isVacantPlot = await isVacant(req.params.id);
+  if (!isVacantPlot) throw new BadRequestError("Plot is not vacant");
   const matchedPlot = await plotModel.findByIdAndDelete(req.params.id);
+  await deleteAllSlots(req.params.id);
   res
     .status(StatusCodes.OK)
     .json({ message: "Plot deleted successfully", matchedPlot });
@@ -41,4 +67,33 @@ export const deletePlotById = async (req, res) => {
 export const getPlotByIdPassed = async (id) => {
   const plot = await plotModel.findById(id);
   return plot;
+};
+
+export const updatePlotSlots = async (type, id) => {
+  const plot = await getPlotByIdPassed(id);
+  await plotModel.findByIdAndUpdate(
+    id,
+    {
+      numUnits: type === "delete" ? plot.numUnits - 1 : plot.numUnits,
+      numOccupied:
+        type === "occupied"
+          ? plot.numOccupied + 1
+          : type === "vacant"
+          ? plot.numOccupied - 1
+          : plot.numOccupied,
+      numVacant: type === "vacant" ? plot.numVacant + 1 : plot.numVacant - 1,
+    },
+    { new: true }
+  );
+};
+
+export const totalOccupancy = async () => {
+  return (await plotModel.find({})).reduce(
+    (occ, cur) => {
+      occ.occupied += cur.numOccupied;
+      occ.vacant += cur.numVacant;
+      return occ;
+    },
+    { occupied: 0, vacant: 0 }
+  );
 };
